// types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
// ...
      properties: {
        Row: {
          address: string
          agent_id: string | null
          amenities: string[]
          area_sqft: number
          bathrooms: number
          bedrooms: number
          city: string
          created_at: string
          description: string | null
          id: string
          images: string[]
          is_featured: boolean
          latitude: number
          listing_type: string
          longitude: number
          neighbourhood: string
          price: number
          price_per: string
          property_type: string
          slug: string
          title: string
          updated_at: string
        }
        Insert: {
          address: string
          agent_id?: string | null
          amenities?: string[]
          area_sqft: number
          bathrooms?: number
          bedrooms?: number
          city?: string
          created_at?: string
          description?: string | null
          id?: string
          images?: string[]
          is_featured?: boolean
          latitude: number
          listing_type: string
          longitude: number
          neighbourhood: string
          price: number
          price_per?: string
          property_type: string
          slug: string
          title: string
          updated_at?: string
        }
        Update: {
          address?: string
          agent_id?: string | null
          amenities?: string[]
          area_sqft?: number
          bathrooms?: number
          bedrooms?: number
          city?: string
          created_at?: string
          description?: string | null
          id?: string
          images?: string[]
          is_featured?: boolean
          latitude?: number
          listing_type?: string
          longitude?: number
          neighbourhood?: string
          price?: number
          price_per?: string
          property_type?: string
          slug?: string
          title?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "properties_agent_id_fkey"
            columns: ["agent_id"]
            isOneToOne: false
            referencedRelation: "agents"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}